rteval: code cleanup and fixes
Changes:

* Closed /dev/null after use on each function opening it.

* Make use of subprocess.wait() calls where appropriate.

* Imported the missing SIGKILL symbol in hackbench.py

* Removed the code snippet below from hackbench.py:runload():

         count = 30
         while count > 0 and p.poll() == None:
             time.sleep(1.0)
             count -= 1
         if p.poll() == None:
             os.kill(p.pid, SIGKILL)

+++ rteval/cyclictest.py
@@ -233,7 +233,9 @@ class Cyclictest(Thread):
                 break
             time.sleep(1.0)
         self.debug("stopping")
-        os.kill(c.pid, signal.SIGINT)
+        if c.poll() == None:
+            os.kill(c.pid, signal.SIGINT)
+        c.wait()
         # now parse the histogram output
         for line in c.stdout:
             if line.startswith('#'): continue
@@ -245,6 +247,7 @@ class Cyclictest(Thread):
         for n in self.data.keys():
             self.data[n].reduce()
         self.finished.set()
+        os.close(null)
 
     def genxml(self, x):
         x.openblock('cyclictest')
